Removed GafferScene::Source
