+++ include/GafferScene/ObjectSourceBase.inl
@@ -138,9 +138,22 @@ const Gaffer::ObjectPlug *ObjectSourceBase<BaseType>::sourcePlug() const
 }
 
 template<typename BaseType>
+void ObjectSourceBase<BaseType>::hashGlobals( const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
+{
+	BaseType::hashGlobals( context, parent, h );
+}
+
+template<typename BaseType>
+void ObjectSourceBase<BaseType>::hashAttributes( const SceneNode::ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
+{
+	BaseType::hashAttributes( path, context, parent, h );
+}
+
+template<typename BaseType>
 void ObjectSourceBase<BaseType>::hashBound( const SceneNode::ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
 	BaseType::hashBound( path, context, parent, h );
+	h.append( &path.front(), path.size() );
 	sourcePlug()->hash( h );
 	if( path.size() == 0 )
 	{
@@ -152,6 +165,7 @@ template<typename BaseType>
 void ObjectSourceBase<BaseType>::hashTransform( const SceneNode::ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
 	BaseType::hashTransform( path, context, parent, h );
+	h.append( &path.front(), path.size() );
 	transformPlug()->hash( h );
 }
 
@@ -159,6 +173,7 @@ template<typename BaseType>
 void ObjectSourceBase<BaseType>::hashObject( const SceneNode::ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
 	BaseType::hashObject( path, context, parent, h );
+	h.append( &path.front(), path.size() );
 	sourcePlug()->hash( h );
 }
 
@@ -166,6 +181,7 @@ template<typename BaseType>
 void ObjectSourceBase<BaseType>::hashChildNames( const SceneNode::ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
 	BaseType::hashChildNames( path, context, parent, h );
+	h.append( &path.front(), path.size() );
 	namePlug()->hash( h );
 }
