+++ src/GafferSceneTest/CompoundObjectSource.cpp
@@ -46,7 +46,7 @@ using namespace GafferSceneTest;
 IE_CORE_DEFINERUNTIMETYPED( CompoundObjectSource )
 
 CompoundObjectSource::CompoundObjectSource( const std::string &name )
-	:	Source( name )
+	:	SceneNode( name )
 {
 	addChild( new ObjectPlug( "in", Plug::In, new CompoundObject() ) );
 }
@@ -67,7 +67,7 @@ const Gaffer::ObjectPlug *CompoundObjectSource::inPlug() const
 
 void CompoundObjectSource::affects( const Plug *input, AffectedPlugsContainer &outputs ) const
 {
-	Source::affects( input, outputs );
+	SceneNode::affects( input, outputs );
 	if( input == inPlug() )
 	{
 		for( ValuePlugIterator it( outPlug() ); it != it.end(); it++ )
@@ -79,37 +79,42 @@ void CompoundObjectSource::affects( const Plug *input, AffectedPlugsContainer &o
 
 void CompoundObjectSource::hashBound( const ScenePath &path, const Gaffer::Context *context, const GafferScene::ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	Source::hashBound( path, context, parent, h );
+	SceneNode::hashBound( path, context, parent, h );
+	h.append( &path.front(), path.size() );
 	inPlug()->hash( h );
 }
 
 void CompoundObjectSource::hashTransform( const ScenePath &path, const Gaffer::Context *context, const GafferScene::ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	Source::hashTransform( path, context, parent, h );
+	SceneNode::hashTransform( path, context, parent, h );
+	h.append( &path.front(), path.size() );
 	inPlug()->hash( h );
 }
 
 void CompoundObjectSource::hashAttributes( const ScenePath &path, const Gaffer::Context *context, const GafferScene::ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	Source::hashAttributes( path, context, parent, h );
+	SceneNode::hashAttributes( path, context, parent, h );
+	h.append( &path.front(), path.size() );
 	inPlug()->hash( h );
 }
 
 void CompoundObjectSource::hashObject( const ScenePath &path, const Gaffer::Context *context, const GafferScene::ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	Source::hashObject( path, context, parent, h );
+	SceneNode::hashObject( path, context, parent, h );
+	h.append( &path.front(), path.size() );
 	inPlug()->hash( h );
 }
 
 void CompoundObjectSource::hashChildNames( const ScenePath &path, const Gaffer::Context *context, const GafferScene::ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	Source::hashChildNames( path, context, parent, h );
+	SceneNode::hashChildNames( path, context, parent, h );
+	h.append( &path.front(), path.size() );
 	inPlug()->hash( h );
 }
 
 void CompoundObjectSource::hashGlobals( const Gaffer::Context *context, const GafferScene::ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	Source::hashGlobals( context, parent, h );
+	SceneNode::hashGlobals( context, parent, h );
 	inPlug()->hash( h );
 }
